backend/requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]
python-multipart

# Document Parsing
pdfplumber
PyPDF2
python-docx

# NLP Processing
nltk
scikit-learn
numpy

# Utilities
python-dotenv
aiofiles

# CORS
starlette

# Gemini AI
google-generativeai

# Database
firebase-admin